Split kiva.fonttools._scanner into two modules
